+++ gantry/commands/run.py
@@ -369,6 +369,12 @@ def run(
     if ref is not None:
         git_ref = ref
 
+    if not util.ref_exists_on_remote(git_ref):
+        raise UnpushedChangesError(
+            f"Current git ref '{git_ref}' does not appear to exist on the remote!\n"
+            "Please push your changes and try again."
+        )
+
     # Initialize Beaker client and validate workspace.
     beaker = util.ensure_workspace(
         workspace=workspace, yes=yes, gh_token_secret=gh_token_secret, public_repo=is_public